[libc] Implem. regparmcall calling convention for execl... funcs.
I also beefed up the dependency handling for the libc build.
