+++ libc/asm/Makefile
@@ -23,4 +23,4 @@ out.a: $(OBJS)
 	$(AR) $(ARFLAGS_SUB) $@ $^
 
 clean:
-	rm -f *.[ao]
+	$(RM) *.[aod]
