+++ libc/getent/Makefile
@@ -14,4 +14,4 @@ out.a: $(OBJS)
 	$(AR) $(ARFLAGS_SUB) $@ $^
 
 clean:
-	rm -f *.[ao]
+	$(RM) *.[aod]
