+++ libc/malloc/Makefile
@@ -35,4 +35,4 @@ out.a: $(OBJS)
 	$(AR) $(ARFLAGS_SUB) $@ $^
 
 clean:
-	$(RM) $(OBJS)
+	$(RM) *.[aod]
