+++ libc/math/Makefile
@@ -60,4 +60,4 @@ $(LIB): $(OBJS)
 	$(AR) $(ARFLAGS_SUB) $@ $^
 
 clean:
-	$(RM) $(OBJS) $(LIB)
+	$(RM) *.[aod]
