+++ libc/misc/Makefile
@@ -45,4 +45,4 @@ $(LIB): $(OBJS)
 	$(AR) $(ARFLAGS_SUB) $@ $^
 
 clean:
-	$(RM) $(OBJS) $(LIB)
+	$(RM) *.[aod]
