+++ libc/stdio/Makefile
@@ -58,4 +58,4 @@ out.a: $(OBJS)
 	$(AR) $(ARFLAGS_SUB) $@ $^
 
 clean:
-	rm -f *.[ao]
+	$(RM) *.[aod]
