+++ libc/system/Makefile
@@ -12,17 +12,20 @@ else
 endif
 
 all: out.a
+.PHONY: all
 
 out.a: syscall.mk out
 
 syscall.mk: $(SYSCALLDAT) $(dir)syscall.awk
 	awk -f $(dir)syscall.awk regparmcall=$(regparmcall) $< > $@
 
-out::
+out: syscall.mk
 	$(MAKE) -f $(dir)out.mk
+.PHONY: out
 
-clean::
+clean:
 	touch syscall.mk
 	$(MAKE) -f $(dir)out.mk clean
 	$(RM) syscall.mk out
 	$(RM) -r syscall
+.PHONY: clean
