--- libc/system/execlp.c
@@ -1,9 +1,0 @@
-#ifdef L_execlp
-#include <unistd.h>
-
-int
-execlp(char *fname, char *arg0, ...)
-{
-   return execvp(fname, &arg0);
-}
-#endif
